Make DMA on_unmask states restorable
Instead of registering an on_unmask listener every time we want a
transfer, and unregistering the listener every time we're done with it,
SB16 now registers a handler during init, and keeps this handler
registered throughout the session. The handler function is moved to SB16's
prototype.

The dma_waiting_transfer flag indicates whether the handler should
actually carry out the transfer.