Pause sb16 when cpu is paused or via pause command
Previously, when the cpu is paused, sb16 continues to perform dma
transfers and the speaker continues to play the audio stored in the
buffer (looping), which is not what we want.

More importantly, when we pause the cpu to save the machine state,
because sb16 was continuing with its dma transfers, the machine might
restore back into a confused and muddled state.