Handle simulatneous kbd and mouse streams
Fixes #182.

This patch handles the following conditions:

- Not all mouse/kbd irq handlers check the status port to see whether
the next port 0x60 read is going to return the mouse byte or kbd byte.
For example, KolibriOS, see:
http://websvn.kolibrios.org/filedetails.php?repname=Kolibri+OS&path=%2Fdrivers%2Fmouse%2Fcommouse.asm

- Not all interrupts may result in reading port 0x60 to clear the
interrupt itself. For example, the mouse irq handler for Windows 1 and
98 appears to skip reading port 0x60 if it finds out that the next byte
to be returned is a kbd byte and not a mouse byte. The mouse irq thus
never gets a chance to get cleared, and subsequent irq_raise do not
actually trigger the handler.

- Between the time:
  ... after a mouse interrupt gets raised, or the OS reads
the status port, and
  ... before the OS reads port 0x60,
more keyboard data may be received, and this race condition will make
the OS read a kbd byte when it is expecting a mouse byte, and vice
versa. Thus, this patch ensures that incoming kbd data will not mess up
the next_byte_is_aux state, and vice versa with incoming mouse data.
Instead, next_byte_is_aux is toggled when the current buffer being read
has been cleared.