Add scratch buffer to codegen and refactor functions to indicate it
Since we can only know an instruction's length (and other characteristics) after
its been decoded, this change allows us to first decode the instruction and
generate the instruction function calls to a scratch buffer, which can then be
copied over to the main (cs) buffer as appropriate.

Tests for the scratch buffer were added too.