Add groundwork for EIP update optimization for non-faulting instructions
Instructions which set the JIT_INSTR_NONFAULTING_FLAG will not update
previous_ip. They will "patch" the last instruction_pointer update instead of
adding to instruction_pointer individually.

This commit also fixes a bug with gen_increment_instruction_pointer not encoding
values to LEB-128; this was okay since previously it was only used for values
less than 16, for which the encoding is the same as the raw value.