codegen: Add 2 local scratch variables for easier inlining
WebAssembly has no "dup" operator (to duplicate the value on the top of
the stack), which means that if a value needs to be used twice, we need
to use a variable. We add 2 general-purpose local scratch variables to
this end.

Note: Users of the variable should be careful about interference with
other users - they may contain junk values, and between functions, the
state of a variable may well change.