cpu: Move tlb_data and page_first_jit_cache_entry to global_pointers
When they're stack-allocated, their addresses may change very easily,
even simply through updates in a header file that cpu.c includes.

This will make the expect-tests fail unnecessarily, simply because these
offsets changed.

In the future, we may change how memory handling works, so this may be a
non-issue then, but for now, it's easier to use a fixed offset for these
arrays.