Optimise gen_safe_write32: Remove jit cache check
The jit cache check (whether a physical page contains code) has been
integrated into the tlb: When code is created or deleted, the tlb is
iterated to set or clear all "has_code" bits. When a new tlb entry is
created, the mapped physical page is checked and the "has_code" bit is
set correspondingly.

The new invariant is:

    (page_entry_points[p] != empty || page_first_jit_cache_entry[p] != empty)
        implies
    (forall v: if(tlb_data[v] == p) tlb_data[v].has_code)