Fix 9p capabilities test: getxattr expects revision 2+ format
Even though some man pages say that linux continues to support
capabilities revision 1, it seems that when cap_inode_getsecurity
handles the retrieval of security.capability, it throws an EINVAL when
the revision number is not 2 nor 3.

At time of writing, the code of interest is at:
line 415 of linux/security/commoncap.c

Also watch out for when it converts revision 3 caps data into revision 2
caps data in certain situations.