Filesystem: Simplify nlinks management across filesystems
Now that hard links aren't allowed across filesystems, it's ok to store
the relevant nlinks information within local inode.

For mountpoints (forwarder inodes that point to a root inode), it is
possible to bump the nlinks of the forwarder during link_under_dir and
unlink_from_dir so that this nlinks counter is modified independently
from the real inode's nlinks counter.

For divert(), it is actually ok to divert forwarder inodes that point to
hardlinked files, so I've relaxed the assertion predicate.