readme: add SkiffOS config link
SkiffOS is a configuration layering system for Buildroot.

It supports re-targeting system image configs to different hardware.

The linked configuration can be combined with other layers to deploy Docker and
other types of applications to the v86 emulator.

The configuration fragments slim down the Buildroot and Kernel systems &
switch off any unnecessary features.

https://github.com/skiffos/SkiffOS/tree/master/configs/browser/v86

Signed-off-by: Christian Stewart <christian@paral.in>