Remove extra_images, add V86.load_image
load_image takes an image file descriptor (like you'd find in a V86
option like `hda` or `cdrom`) and returns a Promise which will resolve
with a full image description (including a `buffer` field which is a
buffer type from buffer.js).

This currently duplicates a bunch of logic from starter's
`continue_init` function, but I hope in the future that could be
refactored out.  The main sticking point is progress reporting, which
would need to do something like add up the total sizes across all
images and increment a counter as new bytes come in, but I wanted to
keep this commit focused.