Try to avoid HTTP compression in load_file (#919)
Web servers might use HTTP compression (Content-Encoding) to send
data, including the image files v86 uses.  Unfortunately, the data
ranges used in AsyncXHRBuffer are in terms of the raw image files and
not the compressed file (where in fact data might need to be pulled
from several ranges and decompressed, and it might be hard to know what
ranges of the compressed file represent what parts of the original
file).

This patch uses two techniques to try and handle this issue:

1. First, we send the nonstandard X-Accept-Encoding header, which some
web servers respect.
2. Second, if we notice a web server has responded to a Range request
with a response with a Content-Encoding header distinct from
"identity" (i.e., with compressed data), we log an error message to
the console.