replaced boolean options.disable_autoscale with number options.scale
options.scale defines the initial value for ScreenAdapter.scale_x and ScreenAdapter.scale_y.

If undefined, options.scale defaults to 1.

If options.scale is set to 0, 2x-upscale, scale_x, scale_y and HiDPI-adaption (window.devicePixelRatio) are all disabled in all three modes, they are in the responsibilty of host applications that define options.scale to be 0.