further relaxed disk geometry detection from image size
Reverted to the original method of expanding any image to its closest fitting geometry if its exact size cannot be matched. This now includes images with a size smaller or equal to 512 bytes which are expanded to the smallest known geometry of 160K.

This is the most robust method, loading an image will now only fail if the given image is larger than any known geometry.