added keyboard layout mapping support to DataKeyboard.send_keypress()
If a key name with a length larger than 1 is passed to send_keypress() then the pyhsical keyboard mapping SCANCODE[] is used to determine the scancode. All key names in the KeyboardEvent.code naming convention have at least a length of 2.

New: If a key name of length 1 is passed to send_keypress() then locale-dependent layout mapping KEYMAPS[] is used to determine the scancode. The raw scancode of the key's first (scancode, modifier) pair is used, modifier is ignored here.

This is a bit closer to the removed V86.keyboard_send_keys(codes, delay) (which was based on deprecated KeyboardEvent.keyCode).