allow to override the construction of XTerm's Terminal object
Made argument "element" in SerialAdapterXtermJS to be either a HTML DOM object (as before) or a callable (new) that returns a pair of a HTML DOM object and a Terminal instance.

As before, if element is a DOM object, a Terminal instance with fixed configuration is created on the element. This change is fully backward-compatible.

If element is a function, the caller creates the Terminal instance instead, with full control over the configuration and access to the rich API of Xterm.